+++ tests/testthat/test-prior-latent-model.R
@@ -27,21 +27,21 @@ test_that("epidist_prior warns or errors for latent model non-uniform priors", {
 
   # Case 1: swindow_raw non-uniform -> Error
   p1 <- mock_prior("normal(0,1)", dpar = "swindow_raw")
-  expect_error(.check_latent_priors(data, p1), "secondary event")
+  expect_error(.check_model_prior(data, p1), "secondary event")
 
   # Case 2: swindow_raw uniform -> No error
   p2 <- mock_prior("uniform(0,1)", dpar = "swindow_raw")
-  expect_no_error(.check_latent_priors(data, p2))
+  expect_no_error(.check_model_prior(data, p2))
 
   # Case 3: pwindow_raw non-uniform -> Warning
   p3 <- mock_prior("normal(0,1)", dpar = "pwindow_raw")
-  expect_warning(.check_latent_priors(data, p3), "primary event")
+  expect_warning(.check_model_prior(data, p3), "primary event")
 
   # Case 4: pwindow_raw uniform -> No warning
   p4 <- mock_prior("uniform(0,1)", dpar = "pwindow_raw")
-  expect_no_warning(.check_latent_priors(data, p4))
+  expect_no_warning(.check_model_prior(data, p4))
 
   # Case 5: mixed
   p5 <- rbind(p1, p3)
-  expect_error(.check_latent_priors(data, p5), "secondary event")
+  expect_error(.check_model_prior(data, p5), "secondary event")
 })